+++ src/app/ReactQueryProvider/useCreateQueryClient.ts
@@ -2,7 +2,7 @@ import { useRef } from "react";
 import { QueryClient, QueryCache, MutationCache } from "@tanstack/react-query";
 import { HttpError } from "@/shared/lib";
 import { useSnackBarStore } from "@/shared/store";
-import { useOverlayStore } from "@/shared/store/overlay";
+import { useOverlayStore } from "@/shared/store";
 import { useRefreshToken } from "./errorHandlers";
 
 export const useCreateQueryClient = () => {
